LastDetails

+++ webforumvue/src/components/ToolBar.vue
@@ -13,7 +13,7 @@
         <v-btn text @click="enableOptions = !enableOptions">{{user.data.email}}</v-btn>
         <template v-if="enableOptions" class="options">
           <v-btn @click="signOut" text>Sign Out</v-btn>
-          <v-btn @click="deleteAccount" text>Delete Account</v-btn>
+          <v-btn @click="deleteOption = true" text>Delete Account</v-btn>
           <v-btn @click="editAccount" text>Edit Account</v-btn>
         </template>
       </template>
@@ -26,6 +26,34 @@
         </v-btn>
       </template>
     </v-toolbar-items>
+    <v-dialog v-model="deleteOption" persistent max-width="600px">
+    <v-card>
+      <v-card-title>
+        <span class="headline">Delete Account</span>
+      </v-card-title>
+      <v-card-text>
+        <v-container>
+          <v-row>
+            <v-col cols="12" md="12">
+              <v-text-field
+                v-model="confirmpassword"
+                :append-icon="showpassword ? 'mdi-eye' : 'mdi-eye-off'"
+                :type="showpassword ? 'text' : 'password'"
+                label="Confirm password to delete account"
+                @click:append="showpassword = !showpassword"
+                prepend-icon="mdi-lock"
+            ></v-text-field>
+            </v-col>
+          </v-row>
+        </v-container>
+      </v-card-text>
+      <v-card-actions style="background-color:#1976d2">
+        <v-spacer></v-spacer>
+        <v-btn color="white" text @click="deleteOption = false">Close</v-btn>
+        <v-btn color="white" text @click="deleteAccount" @hideDeleteOptions="deleteOption=false">Delete</v-btn>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
   </v-toolbar>
 </template>
 
@@ -42,7 +70,11 @@ export default {
   },
   data() {
     return {
-      enableOptions: false
+      confirmpassword: "",
+      showpassword: false,
+      deleteOption: false,
+      enableOptions: false,
+      db: firebase.firestore()
     };
   },
   methods: {
@@ -60,10 +92,40 @@ export default {
         });
     },
     deleteAccount() {
-      var user = firebase.auth().currentUser;
-      user.delete().then(function() {
-          // User deleted.
-        })
+      this.db.collection('entries').get().then((entries) =>
+      {
+        let deletable = true
+        for (let index = 0; index < entries.length && deletable; index++) 
+        {
+          const entry = entries.docs[index];
+          if(entry.data().creator.id === this.user.data.email)
+          {
+            deletable = false
+          }
+        }
+        if(deletable)
+        {
+          let user = firebase.auth().currentUser
+          let userEmail= user.email
+          var credential = firebase.auth.EmailAuthProvider.credential(userEmail, this.confirmpassword)
+          user.reauthenticateWithCredential(credential).then(() => 
+          {
+            let db = this.db
+            user.delete().then(function() 
+            {
+              db.collection('users').doc(userEmail).delete()
+              
+            }).catch(error => 
+            {
+              console.log(error.message)
+            })
+           })
+        }
+        this.deleteOption = false
+      }).catch(error =>
+      {
+        console.log(error.message)
+      })
     },
     editAccount(){
       this.$router.replace(
